+++ packages/typespec-go/test/local/azblob/zz_models.go
@@ -64,13 +64,13 @@ type BlockList struct {
 // BlockLookupList - The Block lookup list.
 type BlockLookupList struct {
 	// The committed blocks
-	Committed [][]byte `xml:"committed"`
+	Committed [][]byte `xml:"Committed"`
 
 	// The latest blocks
-	Latest []*string `xml:"latest"`
+	Latest []*string `xml:"Latest"`
 
 	// The uncommitted blocks
-	Uncommitted [][]byte `xml:"uncommitted"`
+	Uncommitted [][]byte `xml:"Uncommitted"`
 }
 
 // CORSRule - CORS is an HTTP feature that enables a web application running under one domain to access resources in another
@@ -398,13 +398,13 @@ type Metrics struct {
 // PageList - Represents a page list.
 type PageList struct {
 	// The clear ranges.
-	ClearRange []*ClearRange `xml:"clearRange"`
+	ClearRange []*ClearRange `xml:"ClearRange"`
 
 	// The next marker.
 	NextMarker *string `xml:"NextMarker"`
 
 	// The page ranges.
-	PageRange []*PageRange `xml:"pageRange"`
+	PageRange []*PageRange `xml:"PageRange"`
 }
 
 // PageRange - The page range.
@@ -613,7 +613,7 @@ type SignedIdentifier struct {
 // SignedIdentifiers - Represents an array of signed identifiers
 type SignedIdentifiers struct {
 	// REQUIRED; The array of signed identifiers.
-	Items []*SignedIdentifier `xml:"items"`
+	Items []*SignedIdentifier `xml:"SignedIdentifier"`
 }
 
 // StaticWebsite - The properties that enable an account to host a static website