+++ packages/typespec-go/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added support for specific `@clientOption` decorators.
   * `monomorphicResponseFieldName` on operations that return a monomorphic response to customize the response field name.
   * `omitSerdeMethods` on models to skip emitting un/marshalers.
+  * `responseEnvelopeName` on operations to specify the name of the response envelope.
 
 ## 0.10.4 (2026-03-26)